fix exception name
