+++ Lib9c/Action/AlreadyClaimedGiftsException.cs
@@ -18,20 +18,4 @@ namespace Nekoyume.Action
         {
         }
     }
-
-    [Serializable]
-    public class ClaimableGiftsNotAvailableException : Exception
-    {
-        public ClaimableGiftsNotAvailableException()
-        {
-        }
-
-        public ClaimableGiftsNotAvailableException(string msg) : base(msg)
-        {
-        }
-
-        public ClaimableGiftsNotAvailableException(SerializationInfo info, StreamingContext context) : base(info, context)
-        {
-        }
-    }
 }
